Integrated InterfaceDICOMXMLFile into WeaselXMLReader
All funcionality from InterfaceDICOMXMLFile is absorbed into WeaselXMLReader and InterfaceDICOMXMLFile is trashed, This means all DICOM XML functionality is now handled by the same object.

Instead of

import WEASEL.CoreModules.InterfaceDICOMXMLFile as InterfaceDICOMXMLFile
InterfaceDICOMXMLFile.removeOneStudyFromXMLFile(weasel)

we now write

weasel.objXMLReader.removeOneStudyFromXMLFile()

This substitution is made in all modules making use of InterfaceXMLFile.

WeaselXMLReader is enhanced with an optional weasel argument to still allow messaging to the user during or after calculations. If weasel is omitted as in command line operation, the message is printed instead.